+++ R/dateSequence.R
@@ -52,6 +52,8 @@
 #' @export
 #'
 #' @examples
+#' library(MazamaCoreUtils)
+#'
 #' dateSequence("2019-11-01", "2019-11-08", timezone = "America/Los_Angeles")
 #' dateSequence("2019-11-01", "2019-11-07", timezone = "America/Los_Angeles",
 #'              ceilingEnd = TRUE)