+++ R/timeRange.R
@@ -37,7 +37,10 @@
 #' @return A vector of two \code{POSIXct}s.
 #'
 #' @examples
+#' library(MazamaCoreUtils)
+#'
 #' timeRange("2019-01-08 10:12:15", 20190109102030, timezone = "UTC")
+#'
 timeRange <- function(
   starttime = NULL,
   endtime = NULL,