+++ R/timeStamp.R
@@ -22,6 +22,7 @@
 #'
 #' \itemize{
 #'   \item{\code{"ymdhms"}}
+#'   \item{\code{"ymdThms"}}
 #'   \item{\code{"julian"}}
 #'   \item{\code{"clock"}}
 #' }
@@ -36,6 +37,7 @@
 #'   \item{\code{"hour"}}
 #'   \item{\code{"min"}}
 #'   \item{\code{"sec"}}
+#'   \item{\code{"msec"}}
 #' }
 #'
 #' If `style == "julian"` && `unit = "month"``, the timestamp will contain the
@@ -46,6 +48,8 @@
 #' @return A vector of time stamps.
 #'
 #' @examples
+#' library(MazamaCoreUtils)
+#'
 #' datetime <- parseDatetime("2019-01-08 12:30:15", timezone = "UTC")
 #'
 #' timeStamp(datetime, "UTC", unit = "year")
@@ -56,9 +60,11 @@
 #' timeStamp(datetime, "UTC", unit = "hour")
 #' timeStamp(datetime, "UTC", unit = "min")
 #' timeStamp(datetime, "UTC", unit = "sec")
+#' timeStamp(datetime, "UTC", unit = "sec", style = "ymdThms")
 #' timeStamp(datetime, "UTC", unit = "sec", style = "julian")
 #' timeStamp(datetime, "UTC", unit = "sec", style = "clock")
 #' timeStamp(datetime, "America/Los_Angeles", unit = "sec", style = "clock")
+#' timeStamp(datetime, "America/Los_Angeles", unit = "msec", style = "clock")
 #'
 timeStamp <- function(
   datetime = NULL,
@@ -77,10 +83,10 @@ timeStamp <- function(
   if ( !timezone %in% base::OlsonNames() )
     stop(paste0("Timezone '", timezone, "' is not recognized."))
 
-  if ( !unit %in% c("year", "month", "day", "hour", "min", "sec") )
+  if ( !unit %in% c("year", "month", "day", "hour", "min", "sec", "msec") )
     stop(paste0("Unit '", unit, "' is not recognized."))
 
-  if ( !style %in% c("ymdhms", "julian", "clock") )
+  if ( !style %in% c("ymdhms", "ymdThms", "julian", "clock") )
     stop(paste0("Unit '", style, "' is not recognized."))
 
   # ----- Format datetimes -----------------------------------------------------
@@ -94,6 +100,8 @@ timeStamp <- function(
   } else if ( unit == "month" ) {
     if ( style == "ymdhms" )
       format <- "%Y%m"
+    if ( style == "ymdThms" )
+      format <- "%Y%m"
     if ( style == "julian" ) {
       datetime <- lubridate::floor_date(datetime, unit = "month")
       format <- "%Y%j"
@@ -104,6 +112,8 @@ timeStamp <- function(
   } else if ( unit == "day" ) {
     if ( style == "ymdhms" )
       format <- "%Y%m%d"
+    if ( style == "ymdThms" )
+      format <- "%Y%m%d"
     if ( style == "julian" )
       format <- "%Y%j"
     if ( style == "clock" )
@@ -112,6 +122,8 @@ timeStamp <- function(
   } else if ( unit == "hour" ) {
     if ( style == "ymdhms" )
       format <- "%Y%m%d%H"
+    if ( style == "ymdThms" )
+      format <- "%Y%m%dT%H"
     if ( style == "julian" )
       format <- "%Y%j%H"
     if ( style == "clock" )
@@ -120,6 +132,8 @@ timeStamp <- function(
   } else if ( unit == "min" ) {
     if ( style == "ymdhms" )
       format <- "%Y%m%d%H%M"
+    if ( style == "ymdThms" )
+      format <- "%Y%m%dT%H%M"
     if ( style == "julian" )
       format <- "%Y%j%H%M"
     if ( style == "clock" )
@@ -128,10 +142,22 @@ timeStamp <- function(
   } else if ( unit == "sec" ) {
     if ( style == "ymdhms" )
       format <- "%Y%m%d%H%M%S"
+    if ( style == "ymdThms" )
+      format <- "%Y%m%dT%H%M%S"
     if ( style == "julian" )
       format <- "%Y%j%H%M%S"
     if ( style == "clock" )
       format <- "%Y-%m-%dT%H:%M:%S"
+
+  } else if ( unit == "msec" ) {
+    if ( style == "ymdhms" )
+      format <- "%Y%m%d%H%M%OS3"
+    if ( style == "ymdThms" )
+      format <- "%Y%m%dT%H%M%OS3"
+    if ( style == "julian" )
+      format <- "%Y%j%H%M%OS3"
+    if ( style == "clock" )
+      format <- "%Y-%m-%dT%H:%M:%OS3"
 
   }